requirements.python2.txt
# requirements for Python2 <0.4.1 (0.4.0-dev versions still support Python2)

Click==7.0
graphviz==0.10.1
imageio==2.5.0
numpy==1.16.3
Pillow==6.0.0
protobuf==3.7.1
pytest==4.5.0
pytest-cov==2.7.1
PyYAML==5.1
scipy==1.1.0
setproctitle==1.1.10
six==1.12.0
tensorboardX==1.6
torch==1.2.0
torchvision==0.4.0

-e .[vis]
